+++ package.json
@@ -8,7 +8,7 @@
     "compile": "babel-compile -c tc-rules/babel6 src:lib test:.test",
     "prepublish": "npm run compile",
     "pretest": "npm run compile",
-    "test": "mocha .test/api/*_test.js"
+    "test": "mocha .test/*_test.js"
   },
   "repository": {
     "type": "git",
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "aws-sdk-promise": "0.0.0",
-    "babel-runtime": "^6.1.18",
+    "babel-runtime": "^6.3.19",
     "body-parser": "1.13.3",
     "cryptiles": "^2.0.4",
     "debug": "^2.2.0",
@@ -38,11 +38,11 @@
   },
   "devDependencies": {
     "babel-compile": "^1.0.0",
-    "babel-core": "^6.1.21",
+    "babel-core": "^6.3.21",
     "babel-eslint": "^4.1.5",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.1.20",
-    "babel-plugin-transform-runtime": "^6.1.18",
-    "babel-preset-stage-1": "^6.1.18",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.3.16",
+    "babel-plugin-transform-runtime": "^6.3.13",
+    "babel-preset-stage-1": "^6.3.13",
     "eslint": "^1.9.0",
     "mocha": "2.0.1",
     "taskcluster-base": "^0.8.10",
