+++ test/validate_test.js
@@ -1,27 +1,27 @@
-suite("api/schemaPrefix", function() {
+suite("api/validate", function() {
   require('superagent-hawk')(require('superagent'));
   var request         = require('superagent-promise');
   var assert          = require('assert');
   var Promise         = require('promise');
   var mockAuthServer  = require('taskcluster-lib-testing/.test/mockauthserver');
   var base            = require('taskcluster-base');
-  var subject         = require('../../');
+  var subject         = require('../');
   var express         = require('express');
   var path            = require('path');
+
 
   // Create test api
   var api = new subject({
     title:        "Test Api",
-    description:  "Another test api",
-    schemaPrefix: 'http://localhost:4321/'
+    description:  "Another test api"
   });
 
   // Declare a method we can test input with
   api.declare({
-    method:   'get',
+    method:   'post',
     route:    '/test-input',
     name:     'testInput',
-    input:    'test-schema.json',
+    input:    'http://localhost:4321/test-schema.json',
     title:    "Test End-Point",
     description:  "Place we can call to test something",
   }, function(req, res) {
@@ -33,11 +33,49 @@ suite("api/schemaPrefix", function() {
     method:   'get',
     route:    '/test-output',
     name:     'testInput',
-    output:   'test-schema.json',
+    output:   'http://localhost:4321/test-schema.json',
     title:    "Test End-Point",
     description:  "Place we can call to test something",
   }, function(req, res) {
     res.reply({value: 4});
+  });
+
+  // Declare a method we can use to test invalid output
+  api.declare({
+    method:   'get',
+    route:    '/test-invalid-output',
+    name:     'testInput',
+    output:   'http://localhost:4321/test-schema.json',
+    title:    "Test End-Point",
+    description:  "Place we can call to test something",
+  }, function(req, res) {
+    res.reply({value: 12});
+  });
+
+  // Declare a method we can test input validation skipping on
+  api.declare({
+    method:   'post',
+    route:    '/test-skip-input-validation',
+    name:     'testInputSkipInputValidation',
+    input:    'http://localhost:4321/test-schema.json',
+    skipInputValidation: true,
+    title:    "Test End-Point",
+    description:  "Place we can call to test something",
+  }, function(req, res) {
+    res.status(200).send("Hello World");
+  });
+
+  // Declare a method we can test output validation skipping on
+  api.declare({
+    method:   'get',
+    route:    '/test-skip-output-validation',
+    name:     'testOutputSkipInputValidation',
+    output:    'http://localhost:4321/test-schema.json',
+    skipOutputValidation: true,
+    title:    "Test End-Point",
+    description:  "Place we can call to test something",
+  }, function(req, res) {
+    res.reply({value: 12});
   });
 
   // Reference to mock authentication server
@@ -115,7 +153,7 @@ suite("api/schemaPrefix", function() {
   test("input (valid)", function() {
     var url = 'http://localhost:61515/test-input';
     return request
-      .get(url)
+      .post(url)
       .send({value: 5})
       .end()
       .then(function(res) {
@@ -128,7 +166,7 @@ suite("api/schemaPrefix", function() {
   test("input (invalid)", function() {
     var url = 'http://localhost:61515/test-input';
     return request
-      .get(url)
+      .post(url)
       .send({value: 11})
       .end()
       .then(function(res) {
@@ -145,6 +183,66 @@ suite("api/schemaPrefix", function() {
       .then(function(res) {
         assert(res.ok, "Request okay");
         assert(res.body.value === 4, "Got wrong value");
+      });
+  });
+
+  // test invalid output
+  test("output (invalid)", function() {
+    var url = 'http://localhost:61515/test-invalid-output';
+    return request
+      .get(url)
+      .end()
+      .then(function(res) {
+        assert(res.status === 500, "Request wasn't 500");
+      });
+  });
+
+  // test skipping input validation
+  test("skip input validation", function() {
+    var url = 'http://localhost:61515/test-skip-input-validation';
+    return request
+      .post(url)
+      .send({value: 100})
+      .end()
+      .then(function(res) {
+        assert(res.ok, "Request failed");
+        assert(res.text === "Hello World", "Got wrong value");
+      });
+  });
+
+  // test skipping output validation
+  test("skip output validation", function() {
+    var url = 'http://localhost:61515/test-skip-output-validation';
+    return request
+      .get(url)
+      .end()
+      .then(function(res) {
+        assert(res.ok, "Request failed");
+        assert(res.body.value === 12, "Got wrong value");
+      });
+  });
+
+  test("input (correct content-type)", function() {
+    var url = 'http://localhost:61515/test-input';
+    return request
+      .post(url)
+      .send(JSON.stringify({value: 5}))
+      .set('content-type', 'application/json')
+      .end()
+      .then(function(res) {
+        assert(res.status === 200, "Request rejected");
+      });
+  });
+
+  test("input (wrong content-type)", function() {
+    var url = 'http://localhost:61515/test-input';
+    return request
+      .post(url)
+      .send(JSON.stringify({value: 5}))
+      .set('content-type', 'text/x-json')
+      .end()
+      .then(function(res) {
+        assert(res.status === 400, "Request wasn't rejected");
       });
   });
 });
